requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==4.0.0
asyncpg==0.28.0
bcrypt==4.0.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlalchemy2-stubs==0.0.2a35
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2

PyJWT~=2.8.0